fix(session): share one Database per owner across repeat logins (6a4fde3)
`completeLogin` builds a fresh `SessionContainer` every time it runs, and each one
opened its own `Database` on the same file. The previous session's controllers keep
their `Database` alive until they are released, so a launch that logs in repeatedly
ends up with several writer connections competing for one SQLite write lock. The
crash report behind this had five `completeLogin` calls in a single cold launch.

`DatabaseStore` owns the open path — store location, App Group migration, schema
rebuild — and caches the result per owner, so the second login gets the connection
the first one opened. The open path moves out of `SessionAuthenticator` unchanged;
only the caching around it is new.
